+++ errors/codes.go
@@ -1,5 +1,25 @@
 package errors
 
+type Finality string
+
+const (
+	FinalityBusiness                Finality = "Business"
+	FinalityTechnicalRecoverable    Finality = "Technical Recoverable"
+	FinalityTechnicalNonRecoverable Finality = "Technical Non-Recoverable"
+	FinalityAmbiguous               Finality = "Ambiguous"
+)
+
+type Category string
+
+const (
+	CategoryVAL  Category = "VAL"
+	CategoryAUTH Category = "AUTH"
+	CategorySES  Category = "SES"
+	CategorySWI  Category = "SWI"
+	CategoryDB   Category = "DB"
+	CategoryREC  Category = "REC"
+)
+
 type Code string
 
 const (